Add missing step: create template
